e2e: Extract minikube package from clusters
Move all minikube-specific logic (command execution, profile querying,
DNS configuration) into a dedicated e2e/clusters/minikube package behind
a Cluster struct. The clusters package becomes a thin orchestration
layer that manages parallel execution across named clusters.

This separates the "which clusters to manage" concern from the "how to
manage them via minikube" concern, making it easier to swap in a
different cluster provider in the future without changing the
orchestration logic.

The Cluster struct approach (minikube.New(name).Action()) supports
adding per-cluster options later (e.g., driver, memory) without growing
function parameter lists.
